feat: separate packages, apply di into slipp application, code reformat
