+++ nextstep-di/src/main/java/nextstep/di/factory/BeanCreator.java
@@ -4,5 +4,6 @@ import java.util.List;
 
 public interface BeanCreator {
     List<Class<?>> getParams();
+
     Object instantiate(Object... params);
 }
