+++ nextstep-di/src/main/java/nextstep/di/factory/BeanFactory.java
@@ -3,6 +3,10 @@ package nextstep.di.factory;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
+import nextstep.di.factory.exception.ImplClassNotFoundException;
+import nextstep.di.factory.exception.PrimitiveTypeInjectionFailException;
+import nextstep.di.factory.exception.RecursiveFieldException;
+import nextstep.di.factory.exception.UninitializedBeanFactoryException;
 import nextstep.stereotype.Controller;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
