--- nextstep-di/src/main/java/nextstep/di/factory/BeanScanner.java
@@ -1,7 +1,0 @@
-package nextstep.di.factory;
-
-import java.util.Map;
-
-public interface BeanScanner {
-    Map<Class<?>, BeanCreator> scan2(Object... basePackages);
-}
