+++ nextstep-di/src/main/java/nextstep/di/factory/exception/InterfaceCannotInstantiatedException.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory;
+package nextstep.di.factory.exception;
 
 public class InterfaceCannotInstantiatedException extends RuntimeException {
     private static final String ERROR_MSG = "인터페이스는 빈으로 등록할 수 없습니다.";
