+++ nextstep-di/src/main/java/nextstep/di/factory/exception/NoDefaultConstructorException.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory;
+package nextstep.di.factory.exception;
 
 public class NoDefaultConstructorException extends RuntimeException {
     private static final String ERROR_MSG = "기본 생성자가 없는 객체는 생성할 수 없습니다.";
