+++ nextstep-di/src/main/java/nextstep/di/factory/exception/ObjectInstantiationFailException.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory;
+package nextstep.di.factory.exception;
 
 public class ObjectInstantiationFailException extends RuntimeException {
     private static final String ERROR_MSG = "의존 주입 객체를 생성하는 데 실패했습니다.";
