+++ nextstep-di/src/main/java/nextstep/di/factory/exception/PrimitiveTypeInjectionFailException.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory;
+package nextstep.di.factory.exception;
 
 public class PrimitiveTypeInjectionFailException extends RuntimeException {
     private static final String ERROR_MSG = "기본형을 주입할 수 없습니다.";
