+++ nextstep-di/src/main/java/nextstep/di/factory/exception/UninitializedBeanFactoryException.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory;
+package nextstep.di.factory.exception;
 
 public class UninitializedBeanFactoryException extends RuntimeException {
     private static final String ERROR_MSG = "초기화되지 않은 빈 팩토리에 접근할 수 없습니다.";
