+++ nextstep-di/src/test/java/nextstep/di/factory/BeanFactoryTest.java
@@ -2,7 +2,8 @@ package nextstep.di.factory;
 
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
-import nextstep.di.factory.example.*;
+import nextstep.di.factory.example.component.*;
+import nextstep.di.factory.exception.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -12,7 +13,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class BeanFactoryTest {
     private BeanFactory beanFactory;
-    private String basePackages = "nextstep.di.factory.example";
+    private String basePackages = "nextstep.di.factory.example.component";
 
     @BeforeEach
     @SuppressWarnings("unchecked")
