+++ nextstep-di/src/test/java/nextstep/di/factory/example/ComponentScanConfig.java
@@ -3,6 +3,8 @@ package nextstep.di.factory.example;
 import nextstep.annotation.Bean;
 import nextstep.annotation.ComponentScan;
 import nextstep.annotation.Configuration;
+import nextstep.di.factory.example.component.MyQnaService;
+import nextstep.di.factory.example.component.QnaController;
 import org.apache.commons.dbcp2.BasicDataSource;
 
 import javax.sql.DataSource;
