--- nextstep-di/src/test/java/nextstep/di/factory/example/NoImplRepository.java
@@ -1,4 +1,0 @@
-package nextstep.di.factory.example;
-
-public interface NoImplRepository {
-}
