--- nextstep-di/src/test/java/nextstep/di/factory/example/UserRepository.java
@@ -1,5 +1,0 @@
-package nextstep.di.factory.example;
-
-public interface UserRepository {
-
-}
