+++ nextstep-di/src/test/java/nextstep/di/factory/example/component/IntermediateRepository.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory.example;
+package nextstep.di.factory.example.component;
 
 public interface IntermediateRepository extends NoImplRepository {
 }
