+++ nextstep-di/src/test/java/nextstep/di/factory/example/component/MyJdbcTemplate.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory.example;
+package nextstep.di.factory.example.component;
 
 import javax.sql.DataSource;
