+++ nextstep-di/src/test/java/nextstep/di/factory/example/component/NoDefaultCtorController.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory.example;
+package nextstep.di.factory.example.component;
 
 public class NoDefaultCtorController {
     private final int a;
