+++ nextstep-di/src/test/java/nextstep/di/factory/example/component/NoImplRepository.java
@@ -1,0 +1,4 @@
+package nextstep.di.factory.example.component;
+
+public interface NoImplRepository {
+}
