+++ nextstep-di/src/test/java/nextstep/di/factory/example/component/UserRepository.java
@@ -1,0 +1,5 @@
+package nextstep.di.factory.example.component;
+
+public interface UserRepository {
+
+}
